fix(src): mypy-clean types and lint-clean imports across visualization/diagrams
- mypy retarget surfaced 12 findings; all fixed at the annotation level:
  add_axes rect lists -> tuples, Path handling in _resolve_path, float() casts
  on numpy scalars (halos, text positions), pos dict typed dict[str, tuple],
  steps arrays converted with .tolist() so annotate/axvline see python scalars
- ruff autofix removed 30 genuinely unused imports (non-re-export sites);
  discopy availability probe kept with noqa: F401 and explanatory comment
- src/**/AGENTS|README|SKILL.md doc-fleet sync for the affected modules

+++ src/visualization/functor_diagrams.py
@@ -298,7 +298,7 @@ def _draw_category_panel(
             continue
         xs = [p[0] for p in pts]
         ys = [p[1] for p in pts]
-        cx, cy = np.mean(xs), np.mean(ys)
+        cx, cy = float(np.mean(xs)), float(np.mean(ys))
         halo_r = (
             max(
                 max(abs(x - cx) for x in xs),
